requirements.txt
Flask>=2.0.0
Pillow>=9.0.0
Werkzeug>=2.0.0
pypdf>=3.0.0
imageio-ffmpeg>=0.4.0
ffmpeg-python>=0.2.0
